Testpage

+++ application/models.py
@@ -23,18 +23,17 @@ class Tanks(db.Model):
     name = db.Column(db.String(25), nullable=True, unique=True)
     description = db.Column(db.String(100), nullable=True)
 
-
 ########## tests table sql ############################
+
 
 class Tests(db.Model):
     id = db.Column(db.Integer, primary_key=True, nullable=False, unique=True)
     tank_id = db.Column(db.Integer, db.ForeignKey('tanks.id'), nullable=False)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
     date = db.Column(db.DateTime, nullable=False, default=datetime.utcnow)
-    ammonia = db.Column(db.Integer, nullable=False)
-    nitrate = db.Column(db.Integer, nullable=False)
-    nitrite = db.Column(db.Integer, nullable=False)
-
+    ammonia = db.Column(db.Float, nullable=False)
+    nitrate = db.Column(db.Float, nullable=False)
+    nitrite = db.Column(db.Float, nullable=False)
 
 @login_manager.user_loader
 def load_user(id):
